formatted and added a bigger smaller comparison
Formatted the code so it is nice and easy to read.
Added in some if comparison statements which compare then tell the user if their guess is to big or too small.
tested this comparison with the guesses, 10, 15, 19, 16, 18. The answer was 18 and all guesses had the correct output message of if the guess was too big or too small